+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `AnimatedValue` is now a subclass of `EventEmitter`
 - `change` event when the most recent value of an `AnimatedValue` instance changes
 - `AnimatedValue.valuesAreEqual` for checking if `change` should be fired
+- Two optional arguments for the `Subcontext` constructor: `canvas` and `image`
 
 ### Changed
 - begin-step interpolator is now in `AnimatedValue` and no longer in each subclass
@@ -20,6 +21,7 @@
 ### Removed
 - Relative time in addKeys
 - Default `Object.drawSelf`
+- `canvas` as an optional dependency (along with Canvas/ImageConstructor stuff in `Subcontext`)
 
 ## 0.1.0
 ### Added